+++ src/features/products/ui/product-detail/product-analysis/components/analysis-summary.tsx
@@ -5,12 +5,18 @@ import { PiStarFourFill } from "react-icons/pi";
 
 import { cn } from "@/common/lib/utils/ui";
 import { TextReveal } from "@/common/components/ui/animation/text-reveal";
+import { Fullscreen } from "lucide-react";
+import Animate from "@/common/components/ui/animation/animate";
 
 interface AnalysisSummaryProps {
   analysisSummaryText?: string;
+  imageCaptionText?: string;
 }
 
-export function AnalysisSummary({ analysisSummaryText }: AnalysisSummaryProps) {
+export function AnalysisSummary({
+  analysisSummaryText,
+  imageCaptionText,
+}: AnalysisSummaryProps) {
   return (
     <div className="space-y-1">
       <h4 className="text-lg font-semibold flex items-center gap-2">
@@ -36,6 +42,23 @@ export function AnalysisSummary({ analysisSummaryText }: AnalysisSummaryProps) {
           triggerOnView
         />
       </div>
+
+      {imageCaptionText ? (
+        <Animate type="fadeIn" duration={0.5} delay={1} triggerOnView>
+          <div className="mt-5 space-y-2">
+            <h5 className="flex items-center gap-2 text-base font-semibold">
+              <Fullscreen className="w-5 h-5" />
+              Image Caption
+            </h5>
+
+            <div>
+              <p className="text-base font-medium text-foreground">
+                {imageCaptionText}
+              </p>
+            </div>
+          </div>
+        </Animate>
+      ) : null}
     </div>
   );
 }